factor out model storage
