+++ src/main/resources/application.conf
@@ -21,4 +21,15 @@ epflbot {
 
   }
 
+  elasticsearch {
+    host = "0.0.0.0"
+    host = ${?EPFLBOT_ELASTICSEARCH_HOST}
+    port = 9300
+    port = ${?EPFLBOT_ELASTICSEARCH_PORT}
+    clusterName = "epflbot"
+    clusterName = ${?EPFLBOT_ELASTICSEARCH_CLUSTERNAME}
+
+  }
+
+
 }
