+++ src/main/scala/ch/epfl/telegram/Config.scala
@@ -1,6 +1,7 @@
 package ch.epfl.telegram
 
+case class ElasticsearchConfig(host: String, port: Int, clusterName: String)
 case class TequilaConfig(host: String, service: String, request: String, require: String, allows: String)
 case class HttpConfig(interface: String, port: Int)
 
-case class EPFLBotConfig(tequila: TequilaConfig, http: HttpConfig)
+case class EPFLBotConfig(elasticsearch: ElasticsearchConfig, tequila: TequilaConfig, http: HttpConfig)
